+++ scripts/backups.js
@@ -11,7 +11,8 @@ module.exports.World_BAckup = () => {
     if (bds.platform === "bedrock"){
         dir_zip = path.join(bds.bds_dir_bedrock, "worlds") //`${require("../index").}/worlds/`
     } else {
-        const world_name = JSON.parse(java_pro(fs.readFileSync(path.join(bds.bds_dir_java, "server.properties"), "utf8").replaceAll("-", "_"))).level_name
+        let file_ = JSON.stringify(java_pro(fs.readFileSync(path.join(bds.bds_dir_java, "server.properties"), "utf8").replaceAll("-", "_")))
+        let world_name = JSON.parse(file_).level_name
         dir_zip = path.join(bds.bds_dir_java, world_name) //`${require("../index").bds_dir_bedrock}/${world_name}/`
     }
     /**
@@ -46,28 +47,25 @@ module.exports.Drive_backup = () => {
     if (bds.platform === "bedrock"){
         dir_zip = path.join(bds.bds_dir_bedrock, "worlds") //`${require("../index").}/worlds/`
     } else {
-        const world_name = JSON.parse(java_pro(fs.readFileSync(path.join(bds.bds_dir_java, "server.properties"), "utf8").replaceAll("-", "_"))).level_name
+        let file_ = JSON.stringify(java_pro(fs.readFileSync(path.join(bds.bds_dir_java, "server.properties"), "utf8").replaceAll("-", "_"))).split("true").join(true).toString().split("false").join(false).toString()
+        let world_name = JSON.parse(file_).level_name
         dir_zip = path.join(bds.bds_dir_java, world_name) //`${require("../index").bds_dir_bedrock}/${world_name}/`
     }
     /**
      * Before we can start it is good for the server not to have a Corrupted Backup
      * this is only necessary once after the server has started manually
     */
-    if (bds.bds_detect()){bds.stop()}
-    var status_b = true
-    var zip = new AdmZip();
-    zip.addLocalFolder(dir_zip);
-    zip.addZipComment(`Backup zip file in ${today}. \nBackup made to ${process.platform}, Free and open content for all\n\nSirherobrine23© By Bds Maneger.`);
-    zip.writeZip(name);
-    let es = 1;
-    for(es == "0";es++;){
-        if (!(status_b)) break
-    }
     const js_ = {
         "file_path": name,
         "file_name": `bds_backup_World_${today}.zip`,
         "id": undefined
     }
+    console.info("Please wait")
+    if (bds.bds_detect()){bds.stop()}
+    var zip = new AdmZip();
+    zip.addLocalFolder(dir_zip);
+    zip.addZipComment(`Backup zip file in ${today}. \nBackup made to ${process.platform}, Free and open content for all\n\nSirherobrine23© By Bds Maneger.`);
+    zip.writeZip(name);
     return js_
 };